Add separate server sending script to code
Adding an external Python file that will handle sending to the Geosciences server. Currently loss of network can cause the sensor reads to block. This code will eventually be rolled into the main executable itself, but for now has an attached GUI to start/stop sends to the server. When using this, the server options should be disabled in the main Sensors Bridge software.